fixed missing vcb variables

+++ fsInit.h
@@ -5,32 +5,32 @@
 typedef struct{
     //volume name
     //magic number
-    uint64_t blockCount; //number of blocks in volume
-    uint64_t sizeOfBlocks;//size of blocks
-    uint64_t freeBlockCount;//free block count
+    int blockCount; //number of blocks in volume
+    int sizeOfBlocks;//size of blocks
+    int freeBlockCount;//free block count
     //free block pointers
-    uint64_t freeFCBCount;//free FCB count
+    int freeFCBCount;//free FCB count
     //FCB pointers
     //free space list pointer
 
-    uint64_t rdLoc; //root directory pointer/index
-    uint64_t rdBlkCnt;        //blocks taken up by root directory
+    int rdLoc; //root directory pointer/index
+    int rdBlkCnt;        //blocks taken up by root directory
 
     //uint64_t fslLoc;  //free space list pointer/index
-    uint64_t fslBlkCnt;    //blocks taken up by free space list
-    uint64_t fslBytes;
+    int fslBlkCnt;    //blocks taken up by free space list
+    int fslBytes;
 }vCB;
 
 typedef struct{
     unsigned int* freeSpaceBitmap;
-    uint64_t freeSpaceBits;
-    uint64_t location;
-    uint64_t fslBlocksUsed;
+    int freeSpaceBits;
+    int location;
+    int fslBlocksUsed;
 }fSL;
 
 void formatVolume(char* volumeName);
-void initVCB(uint64_t volSize, uint64_t blockSize);
-void initFSL(uint64_t volSize, uint64_t blockSize);
+void initVCB(int volSize, int blockSize);
+void initFSL(int volSize, int blockSize);
 void initRD();
 
 #endif
